+++ example/lib/pages/mini_games/multi_scenario/maps/map_biome_1.dart
@@ -48,11 +48,7 @@ class MapBiome1 extends StatelessWidget {
         moveOnlyMapArea: true,
         zoom: getZoomFromMaxVisibleTile(context, defaultTileSize, 20),
       ),
-      progress: Container(
-        width: double.maxFinite,
-        height: double.maxFinite,
-        color: Colors.black,
-      ),
+      
     );
   }